@@ -18,26 +18,17 @@ namespace Valour.Database.Items.Planets.Members;
 /// <summary>
 /// This represents a user within a planet and is used to represent membership
 /// </summary>
-public class PlanetBan : Item, IPlanetItemAPI<PlanetBan>, INodeSpecific
+public class PlanetBan : PlanetItem, ISharedPlanetBan, INodeSpecific
 {
-    [ForeignKey("Planet_Id")]
-    [JsonIgnore]
-    public virtual Planet Planet { get; set; }
-
-    /// <summary>
-    /// The user that was banned
-    /// </summary>
-    public ulong User_Id { get; set; }
-
-    /// <summary>
-    /// The planet the user was within
-    /// </summary>
-    public ulong Planet_Id { get; set; }
-
     /// <summary>
     /// The member that banned the user
     /// </summary>
     public ulong Banner_Id { get; set; }
+
+    /// <summary>
+    /// The user_id of the target that was banned
+    /// </summary>
+    public ulong Target_Id { get; set; }
 
     /// <summary>
     /// The reason for the ban
@@ -47,12 +38,12 @@ public class PlanetBan : Item, IPlanetItemAPI<PlanetBan>, INodeSpecific
     /// <summary>
     /// The time the ban was placed
     /// </summary>
-    public DateTime Creation_Time { get; set; }
+    public DateTime Time { get; set; }
 
     /// <summary>
-    /// The length of the ban
+    /// The time the ban expires. Null for permanent.
     /// </summary>
-    public uint? Minutes { get; set; }
+    public DateTime? Expires { get; set; }
 
     /// <summary>
     /// The type of this item
@@ -62,93 +53,94 @@ public class PlanetBan : Item, IPlanetItemAPI<PlanetBan>, INodeSpecific
     /// <summary>
     /// True if the ban never expires
     /// </summary>
-    public bool Permanent => Minutes is null;
+    public bool Permanent => Expires == null;
 
     /// <summary>
     /// Creates this item in the database and bans the user
     /// </summary>
-    public async Task CreateAsync(ValourDB db)
+    public override async Task CreateAsync(ValourDB db)
     {
-        PlanetMember target = await db.PlanetMembers.FirstOrDefaultAsync(x => x.User_Id == this.User_Id);
+        PlanetMember target = await db.PlanetMembers.FirstOrDefaultAsync(x => x.Id == Target_Id);
 
-        var roles = db.PlanetRoleMembers.Where(x => x.Member_Id == target.Id);
+        var roles = db.PlanetRoleMembers.Where(x => x.Member_Id == target.Id && x.Planet_Id == Planet_Id);
 
-        db.PlanetRoleMembers.RemoveRange(roles);
+        await db.PlanetRoleMembers.BulkDeleteAsync(roles);
 
         db.PlanetMembers.Remove(target);
 
         await db.AddAsync(this);
         await db.SaveChangesAsync();
 
+        foreach (var role in roles)
+        {
+            PlanetHub.NotifyPlanetItemDelete(role);
+        }
+
+        PlanetHub.NotifyPlanetItemDelete(target);
         PlanetHub.NotifyPlanetItemChange(this);
     }
 
-    /// <summary>
-    /// Success if a member has invite permission
-    /// and that the planet is private to get this 
-    /// invite via the API.
-    /// </summary>
-    public async Task<TaskResult> CanGetAsync(PlanetMember member, ValourDB db)
+    public override async Task<TaskResult> CanGetAsync(PlanetMember member, ValourDB db)
     {
-        if (member is null)
-            return new TaskResult(false, "Member not found.");
+        // Members can get their own ban info
+        if (member.Id == Target_Id)
+            return TaskResult.SuccessResult;
 
-        if (!await member.HasPermissionAsync(PlanetPermissions.Ban, db))
-            return new TaskResult(false, "Member lacks PlanetPermissions.Ban");
+        return await CanBanAsync(member, db);
+    }
+
+    public override async Task<TaskResult> CanDeleteAsync(PlanetMember member, ValourDB db) 
+        => await CanBanAsync(member, db);
+
+    public override async Task<TaskResult> CanUpdateAsync(PlanetMember member, PlanetItem old, ValourDB db)
+    {
+        TaskResult canBan = await CanBanAsync(member, db);
+        if (!canBan.Success)
+            return canBan;
+
+        var oldBan = old as PlanetBan;
+
+        if (this.Target_Id != oldBan.Target_Id)
+            return new TaskResult(false, "You cannot change who was banned");
+
+        if (this.Banner_Id != oldBan.Banner_Id)
+            return new TaskResult(false, "You cannot change who banned the user");
+
+        if (this.Time != oldBan.Time)
+            return new TaskResult(false, "You cannot change the creation time");
 
         return TaskResult.SuccessResult;
     }
 
-    public Task<TaskResult> CanDeleteAsync(PlanetMember member, ValourDB db)
-        => Task.FromResult(TaskResult.SuccessResult);
-
-    public Task<TaskResult> CanUpdateAsync(PlanetMember member, ValourDB db) 
-        => Banner_Id != member.Id
-            ? Task.FromResult(new TaskResult(false, $"The banner is not the same as the auth member"))
-            : Task.FromResult(TaskResult.SuccessResult);
-
-    public async Task<TaskResult> CanCreateAsync(PlanetMember member, ValourDB db)
+    public override async Task<TaskResult> CanCreateAsync(PlanetMember member, ValourDB db)
     {
-        var canGet = await this.CanGetAsync(member, db);
-        if (!canGet.Success)
-            return canGet;
+        TaskResult canBan = await CanBanAsync(member, db);
+        if (!canBan.Success)
+            return canBan;
 
-        if (Banner_Id != member.Id)
-            return new TaskResult(false, $"The banner is not the same as the auth member");
-
-        PlanetMember target = await db.PlanetMembers.FirstOrDefaultAsync(x => x.User_Id == this.User_Id);
+        // Ensure target exists
+        PlanetMember target = await db.PlanetMembers.FirstOrDefaultAsync(x => x.User_Id == Target_Id);
 
         if (target is null)
             return new TaskResult(false, $"Target not found");
 
-        if (Banner_Id != target.Id)
+        if (Banner_Id == member.Id)
             return new TaskResult(false, $"You cannot ban yourself");
 
-
         if (await target.GetAuthorityAsync() >= await member.GetAuthorityAsync())
-            return new TaskResult(false, "You cannot manage roles with higher or same authority than your own.");
+            return new TaskResult(false, "You cannot ban users with higher or same authority than your own.");
+
+        if (Banner_Id != member.Id)
+            return new TaskResult(false, $"The banner is not the same as the auth member");
+
+        // Set time from server
+        Time = DateTime.UtcNow;
 
         return TaskResult.SuccessResult;
     }
-    public Task<TaskResult> ValidateItemAsync(PlanetBan old, ValourDB db)
-    {
-        // This role is new
-        if (old is null)
-        {
-            Creation_Time = DateTime.UtcNow;
-        }
-        else
-        {
-            if (this.User_Id != old.User_Id)
-                return Task.FromResult(new TaskResult(false, "You cannot change who was banned"));
-            if (this.Banner_Id != old.Banner_Id)
-                return Task.FromResult(new TaskResult(false, "You cannot change who banned the user"));
-            if (this.Creation_Time != old.Creation_Time)
-                return Task.FromResult(new TaskResult(false, "You cannot change the creation time"));
-        }
-
-        if (Minutes == 0) Minutes = null;
-
-        return Task.FromResult(TaskResult.SuccessResult);
-    }
+    
+    private static async Task<TaskResult> CanBanAsync(PlanetMember member, ValourDB db) 
+        => !await member.HasPermissionAsync(PlanetPermissions.Ban, db)
+            ? new TaskResult(false, "Member lacks Planet Permission " + PlanetPermissions.Ban.Name)
+            : TaskResult.SuccessResult;
 }
